Creating a git repo for what I've done so far.
I've got a lot of work done on my shitty scraping/parsing code, but it's
still slow going. I'm doing a lot of trial-and-error and running
repeatedly to verify output/lack of crashes. Once I've got it scraping
everything successfully, I'll start actually dumping it into a database
and begin on the merging process.